requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==23.2.0
beautifulsoup4==4.12.3
billiard==4.2.0
black==24.8.0
celery==5.4.0
certifi==2024.7.4
cffi==1.16.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.8
Django==5.0.6
django-encrypted-model-fields==0.6.5
et-xmlfile==1.1.0
flake8==7.1.1
gunicorn==23.0.0
h11==0.14.0
idna==3.7
kombu==5.3.7
mccabe==0.7.0
mypy-extensions==1.0.0
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
prompt_toolkit==3.0.47
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.8
selenium==4.22.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.0
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0
